Create 3. Longest Substring Without Repeating Characters
LeetCode Question 3.
A solution for beginners, which is straightforward, easy to understand, without too many complications and room to optimize once you understand the basic premise of the question. Hope this helps!

Time Complexity: O(n)
Space Complexity: O(min of a,b) for the unordered set. (a, is the upper bound of the space complexity)
Where a: Size of the string
b: Size of the number of characters in the character-set